+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.6
+- Handle more date parsing failures. 
+- Get width, height, depth and weight from items
 ## 0.0.5
 
 - Merged pull request #3 allowing options to be passed into the get_items and get_variations request